+++ cluster/images/kubemark/Dockerfile
@@ -16,4 +16,5 @@
 # kept in sycn with distroless_base definition in the WORKSPACE file.
 FROM gcr.io/distroless/base@sha256:7fa7445dfbebae4f4b7ab0e6ef99276e96075ae42584af6286ba080750d6dfe5
 
-COPY kubemark /kubemark
+COPY hollow-kubelet /hollow-kubelet
+COPY hollow-proxy /hollow-proxy
