+++ cmd/kubemark/hollow-proxy/hollow-proxy.go
@@ -1,0 +1,165 @@
+/*
+Copyright 2021 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package main
+
+import (
+	"errors"
+	goflag "flag"
+	"fmt"
+	"math/rand"
+	"os"
+	"time"
+
+	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
+	"k8s.io/klog/v2"
+
+	v1 "k8s.io/api/core/v1"
+	clientset "k8s.io/client-go/kubernetes"
+	restclient "k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
+	"k8s.io/client-go/tools/record"
+	cliflag "k8s.io/component-base/cli/flag"
+	"k8s.io/component-base/logs"
+	_ "k8s.io/component-base/metrics/prometheus/restclient" // for client metric registration
+	_ "k8s.io/component-base/metrics/prometheus/version"    // for version metric registration
+	"k8s.io/component-base/version"
+	"k8s.io/component-base/version/verflag"
+	"k8s.io/kubernetes/pkg/api/legacyscheme"
+	"k8s.io/kubernetes/pkg/kubemark"
+	fakeiptables "k8s.io/kubernetes/pkg/util/iptables/testing"
+	fakesysctl "k8s.io/kubernetes/pkg/util/sysctl/testing"
+	fakeexec "k8s.io/utils/exec/testing"
+)
+
+type hollowProxyConfig struct {
+	KubeconfigPath       string
+	NodeName             string
+	ContentType          string
+	UseRealProxier       bool
+	ProxierSyncPeriod    time.Duration
+	ProxierMinSyncPeriod time.Duration
+}
+
+func (c *hollowProxyConfig) addFlags(fs *pflag.FlagSet) {
+	fs.StringVar(&c.KubeconfigPath, "kubeconfig", "/kubeconfig/kubeconfig", "Path to kubeconfig file.")
+	fs.StringVar(&c.NodeName, "name", "fake-node", "Name of this Hollow Node.")
+	fs.StringVar(&c.ContentType, "kube-api-content-type", "application/vnd.kubernetes.protobuf", "ContentType of requests sent to apiserver.")
+	fs.BoolVar(&c.UseRealProxier, "use-real-proxier", true, "Set to true if you want to use real proxier inside hollow-proxy.")
+	fs.DurationVar(&c.ProxierSyncPeriod, "proxier-sync-period", 30*time.Second, "Period that proxy rules are refreshed in hollow-proxy.")
+	fs.DurationVar(&c.ProxierMinSyncPeriod, "proxier-min-sync-period", 0, "Minimum period that proxy rules are refreshed in hollow-proxy.")
+}
+
+func (c *hollowProxyConfig) createClientConfigFromFile() (*restclient.Config, error) {
+	clientConfig, err := clientcmd.LoadFromFile(c.KubeconfigPath)
+	if err != nil {
+		return nil, fmt.Errorf("error while loading kubeconfig from file %v: %v", c.KubeconfigPath, err)
+	}
+	config, err := clientcmd.NewDefaultClientConfig(*clientConfig, &clientcmd.ConfigOverrides{}).ClientConfig()
+	if err != nil {
+		return nil, fmt.Errorf("error while creating kubeconfig: %v", err)
+	}
+	config.ContentType = c.ContentType
+	config.QPS = 10
+	config.Burst = 20
+	return config, nil
+}
+
+func main() {
+	rand.Seed(time.Now().UnixNano())
+
+	command := newHollowProxyCommand()
+
+	// TODO: once we switch everything over to Cobra commands, we can go back to calling
+	// cliflag.InitFlags() (by removing its pflag.Parse() call). For now, we have to set the
+	// normalize func and add the go flag set by hand.
+	pflag.CommandLine.SetNormalizeFunc(cliflag.WordSepNormalizeFunc)
+	pflag.CommandLine.AddGoFlagSet(goflag.CommandLine)
+	// cliflag.InitFlags()
+	logs.InitLogs()
+	defer logs.FlushLogs()
+
+	if err := command.Execute(); err != nil {
+		os.Exit(1)
+	}
+}
+
+// newHollowProxyCommand creates a *cobra.Command object with default parameters
+func newHollowProxyCommand() *cobra.Command {
+	s := &hollowProxyConfig{}
+
+	cmd := &cobra.Command{
+		Use:  "hollow-proxy",
+		Long: "hollow-proxy pretends to be an ordinary kube-proxy but using fake iptables, sysctl etc",
+		Run: func(cmd *cobra.Command, args []string) {
+			verflag.PrintAndExitIfRequested()
+			run(s)
+		},
+		Args: func(cmd *cobra.Command, args []string) error {
+			for _, arg := range args {
+				if len(arg) > 0 {
+					return fmt.Errorf("%q does not take any arguments, got %q", cmd.CommandPath(), args)
+				}
+			}
+			return nil
+		},
+	}
+	s.addFlags(cmd.Flags())
+
+	return cmd
+}
+
+func run(config *hollowProxyConfig) {
+	// To help debugging, immediately log version
+	klog.Infof("Version: %+v", version.Get())
+
+	// create a client to communicate with API server.
+	clientConfig, err := config.createClientConfigFromFile()
+	if err != nil {
+		klog.Fatalf("Failed to create a ClientConfig: %v. Exiting.", err)
+	}
+
+	client, err := clientset.NewForConfig(clientConfig)
+	if err != nil {
+		klog.Fatalf("Failed to create API Server client: %v", err)
+	}
+	iptInterface := fakeiptables.NewFake()
+	sysctl := fakesysctl.NewFake()
+	execer := &fakeexec.FakeExec{
+		LookPathFunc: func(_ string) (string, error) { return "", errors.New("fake execer") },
+	}
+	eventBroadcaster := record.NewBroadcaster()
+	recorder := eventBroadcaster.NewRecorder(legacyscheme.Scheme, v1.EventSource{Component: "kube-proxy", Host: config.NodeName})
+
+	hollowProxy, err := kubemark.NewHollowProxyOrDie(
+		config.NodeName,
+		client,
+		client.CoreV1(),
+		iptInterface,
+		sysctl,
+		execer,
+		eventBroadcaster,
+		recorder,
+		config.UseRealProxier,
+		config.ProxierSyncPeriod,
+		config.ProxierMinSyncPeriod,
+	)
+	if err != nil {
+		klog.Fatalf("Failed to create hollowProxy instance: %v", err)
+	}
+	hollowProxy.Run()
+}
